requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
einops==0.8.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
